+++ crates/ic-query/src/ic/model/requests/mod.rs
@@ -13,7 +13,8 @@ pub use canisters::{
     IcCanisterCountRequest, IcCanisterFilters, IcCanisterPageRequest, IcCanisterRequest,
 };
 pub use icrc_analytics::{
-    IcIcrcAnalyticsRequest, IcIcrcTotalSupplyQuery, IcIcrcTotalSupplyRequest,
+    IcIcrcAnalyticsRequest, IcIcrcIndexedCountKind, IcIcrcIndexedCountRequest,
+    IcIcrcTotalSupplyQuery, IcIcrcTotalSupplyRequest,
 };
 pub use metrics::{IcMetricKind, IcMetricQuery, IcMetricRequest};
 pub use network::{IcBoundaryNodeDataCentersRequest, IcDailyStatsQuery, IcDailyStatsRequest};